# Conflicts:
#	Integrations/build.gradle
#	engine/table/src/main/java/io/deephaven/engine/table/impl/select/ConditionFilter.java
#	py/server/deephaven/experimental/__init__.py

+++ web/client-api/src/main/java/io/deephaven/web/client/api/JsColumnStatistics.java
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2016-2024 Deephaven Data Labs and Patent Pending
+// Copyright (c) 2016-2025 Deephaven Data Labs and Patent Pending
 //
 package io.deephaven.web.client.api;
 
@@ -36,7 +36,7 @@ public class JsColumnStatistics {
         /**
          * The number of unique values in the column.
          */
-        UNIQUE_VALUES("UNIQUE VALUES", "long"),
+        UNIQUE_VALUES("UNIQUE VALUES", "int"),
         /**
          * The sum of all data in the column.
          */
